+++ CHANGELOG.md
@@ -2,6 +2,6 @@
 
 All notable changes to the "Serverpod" extension will be documented in this file.
 
-## [Unreleased - 0.0.1-alpha]
+## [Unreleased - 0.0.1]
 
-- Creates serverpod project
+- Creates serverpod project and opens it in the editor.